Got rid of new GCC 4.3.2 warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk/x3270@433 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ Common/readres.c
@@ -108,7 +108,7 @@ validate_and_split_resource(const char *where, const char *arg,
 	else {
 		xs_warning("%s: Invalid resource syntax '%.*s', name must "
 		    "begin with '%s'",
-		    where, sizeof(me_dot)-1, arg, me_dot);
+		    where, (int)(sizeof(me_dot)-1), arg, me_dot);
 		return -1;
 	}
